Separate out LayerCard with default options to sync with map
