+++ app/scripts/components/exploration/components/datasets/dataset-list-item.tsx
@@ -20,7 +20,7 @@ import {
 } from './dataset-list-item-status';
 import { DatasetChart } from './dataset-chart';
 import { getBlockBoundaries, lumpBlocks } from './block-utils';
-import DataLayerCard from './data-layer-card';
+import { DataLayerCardWithSync } from './data-layer-card';
 import {
   DatasetStatus,
   TimelineDatasetSuccess
@@ -32,10 +32,8 @@ import {
 import { useDatasetHover } from '$components/exploration/hooks/use-dataset-hover';
 import {
   useTimelineDatasetAtom,
-  useTimelineDatasetColormap,
   useTimelineDatasetSettings,
   useTimelineDatasetVisibility,
-  useTimelineDatasetColormapScale,
   useAnalysisVariable
 } from '$components/exploration/atoms/hooks';
 import {
@@ -104,10 +102,8 @@ export function DatasetListItem(props: DatasetListItemProps) {
 
   const { isAnalyzing, runAnalysis } = useAnalysisController();
 
-  const [isVisible, setVisible] = useTimelineDatasetVisibility(datasetAtom);
-  const [colorMap, setColorMap] = useTimelineDatasetColormap(datasetAtom);
-  const [colorMapScale, setColorMapScale] =
-    useTimelineDatasetColormapScale(datasetAtom);
+  const [isVisible] = useTimelineDatasetVisibility(datasetAtom);
+
   const [modalLayerInfo, setModalLayerInfo] = useState<LayerInfoModalData>();
   const [, setSetting] = useTimelineDatasetSettings(datasetAtom);
   const setSelectedVariable = useAnalysisVariable(datasetAtom);
@@ -123,16 +119,6 @@ export function DatasetListItem(props: DatasetListItemProps) {
       { throwOnError: false }
     );
   }, [queryClient, datasetId]);
-
-  const onClickLayerInfo = useCallback(() => {
-    const data: LayerInfoModalData = {
-      name: dataset.data.name,
-      description: dataset.data.description,
-      info: dataset.data.info,
-      parentData: dataset.data.parentDataset
-    };
-    setModalLayerInfo(data);
-  }, [dataset]);
 
   const controls = useDragControls();
 
@@ -185,7 +171,6 @@ export function DatasetListItem(props: DatasetListItemProps) {
   const isAnalysisAndSuccess =
     isAnalyzing && dataset.analysis.status === DatasetStatus.SUCCESS;
 
-  const datasetLegend = dataset.data.legend;
   const analysisMetrics = useMemo(
     () => dataset.settings.analysisMetrics ?? [],
     [dataset]
@@ -213,18 +198,7 @@ export function DatasetListItem(props: DatasetListItemProps) {
         <DatasetHeader>
           <DatasetHeaderInner>
             <div style={{ width: '100%' }} onPointerDown={onDragging}>
-              <DataLayerCard
-                dataset={dataset}
-                datasetAtom={datasetAtom}
-                colorMap={colorMap}
-                colorMapScale={colorMapScale}
-                setColorMap={setColorMap}
-                setColorMapScale={setColorMapScale}
-                isVisible={isVisible}
-                setVisible={setVisible}
-                datasetLegend={datasetLegend}
-                onClickLayerInfo={onClickLayerInfo}
-              />
+              <DataLayerCardWithSync dataset={dataset} />
             </div>
             {modalLayerInfo && (
               <LayerInfoModal
@@ -252,7 +226,8 @@ export function DatasetListItem(props: DatasetListItemProps) {
               {isAnalysisAndLoading && (
                 <DatasetTrackLoading
                   message={
-                    dataset.analysis.meta.total === undefined || dataset.analysis.meta.loaded === undefined
+                    dataset.analysis.meta.total === undefined ||
+                    dataset.analysis.meta.loaded === undefined
                       ? 'Fetching item information'
                       : `${dataset.analysis.meta.loaded} of ${dataset.analysis.meta.total} items loaded`
                   }
